implement multi-category services
